backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
requests
beautifulsoup4
pydantic
python-dotenv
langchain-core
langchain-community
langchain-google-genai
# Optional: gemini/google client will be needed if you use real Gemini API

# For dev/testing
pytest
